Enum ethers::addressbook::Chain

#[repr(u64)]
pub enum Chain {

    Mainnet,
    Morden,
    Ropsten,
    Rinkeby,
    Goerli,
    Kovan,
    Holesky,
    Sepolia,
    Optimism,
    OptimismKovan,
    OptimismGoerli,
    Arbitrum,
    ArbitrumTestnet,
    ArbitrumGoerli,
    ArbitrumNova,
    Cronos,
    CronosTestnet,
    Rsk,
    BinanceSmartChain,
    BinanceSmartChainTestnet,
    Poa,
    Sokol,
    ScrollAlphaTestnet,
    Metis,
    XDai,
    Polygon,
    PolygonMumbai,
    PolygonZkEvm,
    PolygonZkEvmTestnet,
    Fantom,
    FantomTestnet,
    Moonbeam,
    MoonbeamDev,
    Moonriver,
    Moonbase,
    Dev,
    AnvilHardhat,
    Evmos,
    EvmosTestnet,
    Chiado,
    Oasis,
    Emerald,
    EmeraldTestnet,
    FilecoinMainnet,
    FilecoinHyperspaceTestnet,
    Avalanche,
    AvalancheFuji,
    Celo,
    CeloAlfajores,
    CeloBaklava,
    Aurora,
    AuroraTestnet,
    Canto,
    CantoTestnet,
    Boba,
    Base,
    BaseGoerli,
    Linea,
    LineaTestnet,
    ZkSync,
    ZkSyncTestnet,

}

An Ethereum EIP-155 chain.

Variants

Mainnet

Morden

Ropsten

Rinkeby

Goerli

Kovan

Holesky

Sepolia

Optimism

OptimismKovan

OptimismGoerli

Arbitrum

ArbitrumTestnet

ArbitrumGoerli

ArbitrumNova

Cronos

CronosTestnet

Rsk

BinanceSmartChain

BinanceSmartChainTestnet

Poa

Sokol

ScrollAlphaTestnet

Metis

XDai

Polygon

PolygonMumbai

PolygonZkEvm

PolygonZkEvmTestnet

Fantom

FantomTestnet

Moonbeam

MoonbeamDev

Moonriver

Moonbase

Dev

AnvilHardhat

Evmos

EvmosTestnet

Chiado

Oasis

Emerald

EmeraldTestnet

FilecoinMainnet

FilecoinHyperspaceTestnet

Avalanche

AvalancheFuji

Celo

CeloAlfajores

CeloBaklava

Aurora

AuroraTestnet

Canto

CantoTestnet

Boba

Base

BaseGoerli

Linea

LineaTestnet

ZkSync

ZkSyncTestnet

Implementations

impl Chain

pub const fn average_blocktime_hint(&self) -> Option<Duration>

Returns the chain’s average blocktime, if applicable.

It can be beneficial to know the average blocktime to adjust the polling of an HTTP provider for example.

Note: this is not an accurate average, but is rather a sensible default derived from blocktime charts such as Etherscan’s or Polygonscan’s.

Examples

use ethers_core::types::Chain;
use std::time::Duration;

assert_eq!(
    Chain::Mainnet.average_blocktime_hint(),
    Some(Duration::from_millis(12_000)),
);
assert_eq!(
    Chain::Optimism.average_blocktime_hint(),
    Some(Duration::from_millis(2_000)),
);

pub const fn is_legacy(&self) -> bool

Returns whether the chain implements EIP-1559 (with the type 2 EIP-2718 transaction type).

Examples

use ethers_core::types::Chain;

assert!(!Chain::Mainnet.is_legacy());
assert!(Chain::Celo.is_legacy());

pub const fn supports_push0(&self) -> bool

Returns whether the chain supports the PUSH0 opcode or not.

For more information, see EIP-3855: <https://eips.ethereum.org/EIPS/eip-3855>

pub const fn etherscan_urls(&self) -> Option<(&'static str, &'static str)>

Returns the chain’s blockchain explorer and its API (Etherscan and Etherscan-like) URLs.

Returns (API_URL, BASE_URL)

Examples

use ethers_core::types::Chain;

assert_eq!(
    Chain::Mainnet.etherscan_urls(),
    Some(("https://api.etherscan.io/api", "https://etherscan.io"))
);
assert_eq!(
    Chain::Avalanche.etherscan_urls(),
    Some(("https://api.snowtrace.io/api", "https://snowtrace.io"))
);
assert_eq!(Chain::AnvilHardhat.etherscan_urls(), None);

pub const fn etherscan_api_key_name(&self) -> Option<&'static str>

Returns the chain’s blockchain explorer’s API key environment variable’s default name.

Examples

use ethers_core::types::Chain;

assert_eq!(Chain::Mainnet.etherscan_api_key_name(), Some("ETHERSCAN_API_KEY"));
assert_eq!(Chain::AnvilHardhat.etherscan_api_key_name(), None);

pub fn etherscan_api_key(&self) -> Option<String>

Returns the chain’s blockchain explorer’s API key, from the environment variable with the name specified in etherscan_api_key_name.

Examples

use ethers_core::types::Chain;

let chain = Chain::Mainnet;
std::env::set_var(chain.etherscan_api_key_name().unwrap(), "KEY");
assert_eq!(chain.etherscan_api_key().as_deref(), Some("KEY"));