Trait ethers::contract::EthEvent

pub trait EthEvent: Detokenize + Send + Sync {
    // Required methods
    fn name() -> Cow<'static, str>;
    fn signature() -> H256;
    fn abi_signature() -> Cow<'static, str>;
    fn decode_log(log: &RawLog) -> Result<Self, Error>
       where Self: Sized;
    fn is_anonymous() -> bool;

    // Provided method
    fn new<B, M>(filter: Filter, provider: B) -> Event<B, M, Self>
       where Self: Sized,
             B: Borrow<M>,
             M: Middleware { ... }
}

A trait for implementing event bindings

Required Methods

fn name() -> Cow<'static, str>

The name of the event this type represents

fn signature() -> H256

Retrieves the signature for the event this data corresponds to. This signature is the Keccak-256 hash of the ABI signature of this event.

fn abi_signature() -> Cow<'static, str>

Retrieves the ABI signature for the event this data corresponds to.

fn decode_log(log: &RawLog) -> Result<Self, Error>where Self: Sized,

Decodes an Ethereum RawLog into an instance of the type.

fn is_anonymous() -> bool

Returns true if this is an anonymous event

Provided Methods

fn new<B, M>(filter: Filter, provider: B) -> Event<B, M, Self>where Self: Sized, B: Borrow<M>, M: Middleware,

Returns an Event builder for the ethereum event represented by this types ABI signature.