ethers::core::abi::ethabi
Trait ParseLog

pub trait ParseLog {
    type Log;

    // Required method
    fn parse_log(&self, log: RawLog) -> Result<Self::Log, Error>;
}

trait common to things (events) that have an associated Log type that can be parsed from a RawLog

Required Associated Types

type Log

the associated Log type that can be parsed from a RawLog by calling parse_log

Required Methods

fn parse_log(&self, log: RawLog) -> Result<Self::Log, Error>

parse the associated Log type from a RawLog