ethers::core::k256::ecdsa::signature::digest
Crate crypto_common

Available on crate feature ecdsa-core only.

Common cryptographic traits.

Structs

InvalidLength

The error type returned when key and/or IV used in the KeyInit, KeyIvInit, and InnerIvInit slice-based methods had an invalid length.

Traits

AlgorithmName

Trait which stores algorithm name constant, used in Debug implementations.

BlockSizeUser

Types which process data in blocks.

InnerInit

Types which can be initialized from another type (usually block ciphers).

InnerIvInit

Types which can be initialized from another type and additional initialization vector/nonce.

InnerUser

Types which use another type for initialization.

IvSizeUser

Types which use initialization vector (nonce) for initialization.

KeyInit

Types which can be initialized from key.

KeyIvInit

Types which can be initialized from key and initialization vector (nonce).

KeySizeUser

Types which use key for initialization.

OutputSizeUser

Types which return data with the given size.

ParBlocksSizeUser

Types which can process blocks in parallel.

Reset

Resettable types.

Type Aliases

Block

Block on which BlockSizeUser implementors operate.

Iv

Initialization vector (nonce) used by IvSizeUser implementors.

Key

Key used by KeySizeUser implementors.

Output

Output array of OutputSizeUser implementors.

ParBlocks

Parallel blocks on which ParBlocksSizeUser implementors operate.