Module ethers::core::k256::elliptic_curve::bigint::modular

Implements modular arithmetic for constant moduli.

Modules

constant_mod

Implements Residues, supporting modular arithmetic with a constant modulus.

runtime_mod

Implements DynResidues, supporting modular arithmetic with a modulus set at runtime.

Traits

Retrieve

A generalization for numbers kept in optimized representations (e.g. Montgomery) that can be converted back to the original form.

Functions

montgomery_reduction

Algorithm 14.32 in Handbook of Applied Cryptography https://cacr.uwaterloo.ca/hac/about/chap14.pdf