Trait ethers::core::k256::elliptic_curve::bigint::modular::Retrieve

pub trait Retrieve {
    type Output;

    // Required method
    fn retrieve(&self) -> Self::Output;
}

A generalization for numbers kept in optimized representations (e.g. Montgomery) that can be converted back to the original form.

Required Associated Types

type Output

The original type.

Required Methods

fn retrieve(&self) -> Self::Output

Convert the number back from the optimized representation.