Trait ethers::core::k256::elliptic_curve::bigint::prelude::Concat

pub trait Concat: ConcatMixed<Self, MixedOutput = Self::Output> {
    type Output;

    // Provided method
    fn concat(&self, lo: &Self) -> Self::Output { ... }
}

Concatenate two numbers into a “wide” double-width value, using the lo value as the least significant value.

Required Associated Types

type Output

Concatenated output: twice the width of Self.

Provided Methods

fn concat(&self, lo: &Self) -> Self::Output

Concatenate the two halves, with self as most significant and lo as the least significant.