ethers::core::k256::elliptic_curve::point
Trait DecompressPoint

pub trait DecompressPoint<C>: Sized
where
    C: Curve,
{
    // Required method
    fn decompress(
        x: &GenericArray<u8, <C as Curve>::FieldBytesSize>,
        y_is_odd: Choice,
    ) -> CtOption<Self>;
}

Decompress an elliptic curve point.

Point decompression recovers an original curve point from its x-coordinate and a boolean flag indicating whether or not the y-coordinate is odd.

Required Methods

fn decompress( x: &GenericArray<u8, <C as Curve>::FieldBytesSize>, y_is_odd: Choice, ) -> CtOption<Self>

Attempt to decompress an elliptic curve point.

Object Safety

This trait is not object safe.