Trait ethers::core::k256::elliptic_curve::BatchNormalize

pub trait BatchNormalize<Points>: Curve
where
    Points: ?Sized,
{
    type Output: AsRef<[Self::AffineRepr]>;

    // Required method
    fn batch_normalize(points: &Points) -> Self::Output;
}

Normalize point(s) in projective representation by converting them to their affine ones.

Required Associated Types

type Output: AsRef<[Self::AffineRepr]>

The output of the batch normalization; a container of affine points.

Required Methods

fn batch_normalize(points: &Points) -> Self::Output

Perform a batched conversion to affine representation on a sequence of projective points at an amortized cost that should be practically as efficient as a single conversion. Internally, implementors should rely upon InvertBatch.

Object Safety

This trait is not object safe.