ethers::core::types
Module spoof

Provides types and methods for constructing an eth_call state override set

Structs

Account

The state elements to override for a particular account.

State

A wrapper type that holds a complete state override set.

Enums

Storage

Wraps a map from storage slot to the overriden value.

Functions

balance

Returns a state override set with a single element setting the balance of the address.

code

Returns a state override set with a single element setting the code at the address.

nonce

Returns a state override set with a single element setting the nonce of the address.

state

Returns an empty state override set.

storage

Returns a state override set with a single element setting the storage at the given address and key.