ethers::core::types::transaction
Module eip712

Structs

EIP712Domain

Eip712 Domain attributes used in determining the domain separator; Unused fields are left out of the struct type.

EIP712WithDomain

Eip712DomainType

Represents the name and type pair

TypedData

Represents the EIP-712 typed data object.

Enums

Eip712Error

An EIP-712 error.

Constants

EIP712_DOMAIN_TYPE_HASH

Pre-computed value of the following expression:

EIP712_DOMAIN_TYPE_HASH_WITH_SALT

Pre-computed value of the following expression:

Traits

Eip712

Helper methods for computing the typed data hash used in eth_signTypedData.

Functions

encode_data

Encodes an object by encoding and concatenating each of its members.

encode_eip712_type

Parse token into Eip712 compliant ABI encoding

encode_field

Encode a single field.

encode_type

Encodes the type of an object by encoding a comma delimited list of its members.

hash_struct

Hashes an object

hash_type

Returns the hashed encoded type of primary_type

make_type_hash

Convert hash map of field names and types into a type hash corresponding to enc types;

Type Aliases

Types

Custom types for TypedData