ethers::core::utils::__serde_json::de
Trait Read

pub trait Read<'de>: Sealed { }

Trait used by the deserializer for iterating over input. This is manually “specialized” for iterating over &[u8]. Once feature(specialization) is stable we can use actual specialization.

This trait is sealed and cannot be implemented for types outside of serde_json.

Object Safety

This trait is not object safe.