ethers::etherscan
Module account

Available on crate feature etherscan only.

Structs

AccountBalance

The raw response from the balance-related API endpoints

BeaconWithdrawalTransaction

The raw response from the beacon wihtdrawal transaction list API endpoint

ERC20TokenTransferEvent

The raw response from the ERC20 transfer list API endpoint

ERC721TokenTransferEvent

The raw response from the ERC721 transfer list API endpoint

ERC1155TokenTransferEvent

The raw response from the ERC1155 transfer list API endpoint

InternalTransaction

The raw response from the internal transaction list API endpoint

MinedBlock

The raw response from the mined blocks API endpoint

NormalTransaction

The raw response from the transaction list API endpoint

TxListParams

Common optional arguments for the transaction or event list API endpoints

Enums

BlockType

The pre-defined block type for retrieving mined blocks

GenesisOption

Possible values for some field responses.

InternalTxQueryOption

Options for querying internal transactions

Sort

The list sorting preference

Tag

The pre-defined block parameter for balance API endpoints

TokenQueryOption

Options for querying ERC20 or ERC721 token transfers