ethers::middleware::contract
Type Alias ContractCall

pub type ContractCall<M, D> = FunctionCall<Arc<M>, M, D>;

ContractCall is a FunctionCall object with an std::sync::Arc middleware. This type alias exists to preserve backwards compatibility with less-abstract Contracts.

For full usage docs, see FunctionCall.

Aliased Type

struct ContractCall<M, D> {
    pub tx: TypedTransaction,
    pub function: Function,
    pub block: Option<BlockId>,
    /* private fields */
}

Fields
tx: TypedTransaction

The raw transaction object

function: Function

The ABI of the function being called

block: Option<BlockId>

Optional block number to be used when calculating the transaction’s gas and nonce