ethers::middleware
Module gas_escalator

The Gas Escalator middleware is used to re-broadcast transactions with an increasing gas price to guarantee their timely inclusion.

Structs

EscalationTaskNon-WebAssembly

GasEscalatorMiddleware

A Gas escalator allows bumping transactions’ gas price to avoid getting them stuck in the memory pool.

GeometricGasPrice

Geometrically increasing gas price.

LinearGasPrice

Linearly increasing gas price.

Enums

Frequency

The frequency at which transactions will be bumped

GasEscalatorError

Error thrown when the GasEscalator interacts with the blockchain

Traits

GasEscalator

Trait for fetching updated gas prices after a transaction has been first broadcast