ethers::middleware
Module gas_oracle

The gas oracle middleware is used to get the gas price from a list of gas oracles instead of using eth_gasPrice. For usage examples, refer to the GasOracle trait.

Modules

blocknative

cache

eth_gas_station

etherchain

etherscanetherscan

gas_now

median

middleware

polygon

provider_oracle

Structs

BlockNative

A client over HTTP for the BlockNative gas tracker API that implements the GasOracle trait.

Cache

EthGasStationDeprecated

A client over HTTP for the EthGasStation gas tracker API that implements the GasOracle trait.

Etherchain

A client over HTTP for the Etherchain gas tracker API that implements the GasOracle trait.

Etherscan

A client over HTTP for the Etherscan gas tracker API that implements the GasOracle trait

GasNow

A client over HTTP for the beaconcha.in GasNow gas tracker API that implements the GasOracle trait.

GasOracleMiddleware

Middleware used for fetching gas prices over an API instead of eth_gasPrice.

Median

Polygon

The Polygon gas station API Queries over HTTP and implements the GasOracle trait.

ProviderOracle

Gas oracle from a Middleware implementation such as an Ethereum RPC provider.

Enums

GasCategory

Generic GasOracle gas price categories.

GasOracleError

Error thrown by a GasOracle.

MiddlewareError

Traits

GasOracle

An Ethereum gas price oracle.

Type Aliases

Result