Module ethers::providers::admin

Types for the admin api

Structs

EthInfo

Represents a short summary of the eth sub-protocol metadata known about a connected peer

EthProtocolInfo

Represents a short summary of the eth sub-protocol metadata known about the host peer.

NodeInfo

This includes general information about a running node, spanning networking and protocol details.

PeerInfo

Represents a short summary of information known about a connected peer.

PeerNetworkInfo

Represents networking related information about the peer, including details about whether or not it is inbound, trusted, or static.

PeerProtocolInfo

Represents the protocols that a peer supports.

Ports

Represents a node’s discovery and listener ports.

ProtocolInfo

Represents protocols that the connected RPC node supports.

SnapInfo

Represents a short summary of the snap sub-protocol metadata known about a connected peer.

SnapProtocolInfo

Represents a short summary of the host’s snap sub-protocol metadata.

Enums

EthPeerInfo

Can contain either eth protocol info or a string “handshake”, which geth uses if the peer is still completing the handshake for the protocol.

SnapPeerInfo

Can contain either snap protocol info or a string “handshake”, which geth uses if the peer is still completing the handshake for the protocol.