Struct ethers::providers::PendingTransaction

pub struct PendingTransaction<'a, P> { /* private fields */ }

A pending transaction is a transaction which has been submitted but is not yet mined. await’ing on a pending transaction will resolve to a transaction receipt once the transaction has enough confirmations. The default number of confirmations is 1, but may be adjusted with the confirmations method. If the transaction does not have enough confirmations or is not mined, the future will stay in the pending state.

Example

use ethers_core::types::TransactionRequest;

let tx = TransactionRequest::new().to(to).value(1000).from(from);
let receipt = client
    .send_transaction(tx, None)
    .await?                           // PendingTransaction<_>
    .log_msg("Pending transfer hash") // print pending tx hash with message
    .await?;                          // Result<Option<TransactionReceipt>, _>

Implementations

impl<'a, P> PendingTransaction<'a, P>where P: JsonRpcClient,

pub fn new( tx_hash: H256, provider: &'a Provider<P> ) -> PendingTransaction<'a, P>

Creates a new pending transaction poller from a hash and a provider

pub fn provider(&self) -> Provider<P>where P: Clone,

Returns the Provider associated with the pending transaction

pub fn tx_hash(&self) -> H256

Returns the transaction hash of the pending transaction

pub fn confirmations(self, confs: usize) -> PendingTransaction<'a, P>

Sets the number of confirmations for the pending transaction to resolve to a receipt

pub fn interval<T>(self, duration: T) -> PendingTransaction<'a, P> where T: Into<Duration>,

Sets the polling interval

pub fn retries(self, retries: usize) -> PendingTransaction<'a, P>

Set retries

impl<'a, P> PendingTransaction<'a, P>

pub fn inspect<F>(self, f: F) -> PendingTransaction<'a, P> where F: FnMut(&PendingTransaction<'a, P>),

Allows inspecting the content of a pending transaction in a builder-like way to avoid more verbose calls, e.g.: let mined = token.transfer(recipient, amt).send().await?.inspect(|tx| println!(".{}", *tx)).await?;

pub fn log_msg<S>(self, msg: S) -> PendingTransaction<'a, P> where S: Display,

Logs the pending transaction hash along with a custom message before it.

pub fn log(self) -> PendingTransaction<'a, P>

Logs the pending transaction’s hash