Trait ethers::providers::RawCall

pub trait RawCall<'a> {
    // Required methods
    fn block(self, id: BlockId) -> Self;
    fn state(self, state: &'a State) -> Self;

    // Provided method
    fn map<F>(self, f: F) -> Map<Self, F> 
       where Self: Sized { ... }
}

Provides methods for overriding parameters to the eth_call rpc method

Required Methods

fn block(self, id: BlockId) -> Self

Sets the block number to execute against

fn state(self, state: &'a State) -> Self

Sets the state override set. Note that not all client implementations will support this as a parameter.

Provided Methods

fn map<F>(self, f: F) -> Map<Self, F>
where Self: Sized,

Maps a closure f over the result of .awaiting this call

Object Safety

This trait is not object safe.