ethers::signers
Crate coins_bip39

The bip39 crate is heavily inspired by and reuses code from Wagyu under the MIT license. The difference being, the underlying extended private keys are generated using the bip32 crate, that depends on k256 instead of libsecp256k1.

Modules

mnemonic

Mnemonic phrases

wordlist

Wordlists

Structs

ChineseSimplified

The ChineseSimplified wordlist that implements the Wordlist trait.

ChineseTraditional

The ChineseTraditional wordlist that implements the Wordlist trait.

Czech

The Czech wordlist that implements the Wordlist trait.

English

The English wordlist that implements the Wordlist trait.

French

The French wordlist that implements the Wordlist trait.

Italian

The Italian wordlist that implements the Wordlist trait.

Japanese

The Japanese wordlist that implements the Wordlist trait.

Korean

The Korean wordlist that implements the Wordlist trait.

Mnemonic

Mnemonic represents entropy that can be represented as a phrase. A mnemonic can be used to deterministically generate an extended private key or derive its child keys.

Portuguese

The Portuguese wordlist that implements the Wordlist trait.

Spanish

The Spanish wordlist that implements the Wordlist trait.

Enums

Entropy

Holds valid entropy lengths for a mnemonic

MnemonicError

The error type returned while interacting with mnemonics.

WordlistError

The error type returned while interacting with wordists.

Traits

Wordlist

The Wordlist trait that every language’s wordlist must implement.