Enum ethers::signers::coins_bip39::mnemonic::Entropy

pub enum Entropy {
    Sixteen([u8; 16]),
    Twenty([u8; 20]),
    TwentyFour([u8; 24]),
    TwentyEight([u8; 28]),
    ThirtyTwo([u8; 32]),
}

Holds valid entropy lengths for a mnemonic

Variants

Sixteen([u8; 16])

Sixteen bytes of entropy

Twenty([u8; 20])

Twenty bytes of entropy

TwentyFour([u8; 24])

TwentyFour bytes of entropy

TwentyEight([u8; 28])

TwentyEight bytes of entropy

ThirtyTwo([u8; 32])

ThirtyTwo bytes of entropy

Implementations

impl Entropy

pub fn from_slice(buf: impl AsRef<[u8]>) -> Result<Entropy, MnemonicError>

Attempts to instantiate Entropy from a slice. Fails if the slice is not a valid entropy length

pub fn from_rng<R>(bytes: usize, rng: &mut R) -> Result<Entropy, MnemonicError>where R: Rng,

Instantiates new entropy from an RNG. Fails if the specified bytes is not a valid entropy length

pub const fn words(&self) -> usize

Computes the number of words in the mnemonic

pub const fn len(&self) -> usize

Returns the length of the entropy array