Struct ethers::solc::AllowedLibPaths

pub struct AllowedLibPaths(/* private fields */);

Helper struct for serializing --allow-paths arguments to Solc

From the Solc docs: For security reasons the compiler has restrictions on what directories it can access. Directories of source files specified on the command line and target paths of remappings are automatically allowed to be accessed by the file reader, but everything else is rejected by default. Additional paths (and their subdirectories) can be allowed via the –allow-paths /sample/path,/another/sample/path switch. Everything inside the path specified via –base-path is always allowed.

Implementations

impl AllowedLibPaths

pub fn args(&self) -> Option<[String; 2]>

Returns the Command arguments for this type

--allow-paths takes a single value: all comma separated paths

pub fn paths(&self) -> impl Iterator<Item = &PathBuf>

Returns all paths that exist