Crate ethers::utils::hex

This crate provides a fast conversion of byte arrays to hexadecimal strings, both at compile time, and at run time.

Extends the hex crate’s implementation with const-eval, a const-generics formatting buffer, similar to itoa’s, and more.

Version requirement: rustc 1.64+

Modules

hex

Encoding and decoding hex strings.

Structs

Buffer

A correctly sized stack allocation for the formatted bytes to be written into.

Enums

FromHexError

The error type for decoding a hex string into Vec<u8> or [u8; N].

Constants

HEX_CHARS_LOWER

The table of lowercase characters used for hex encoding.

HEX_CHARS_UPPER

The table of uppercase characters used for hex encoding.

HEX_DECODE_LUT

The lookup table of hex byte to value, used for hex decoding.

Traits

FromHex

Types that can be decoded from a hex string.

ToHex

Encoding values as hex string.

Functions

const_encode

Encodes input as a hex string into a Buffer.

decode

Decodes a hex string into raw bytes.

decode_to_slice

Decode a hex string into a mutable bytes slice.

encode

Encodes data as a hex string using lowercase characters.

encode_prefixed

Encodes data as a prefixed hex string using lowercase characters.

encode_to_slice

Encodes input as a hex string using lowercase characters into a mutable slice of bytes output.

encode_to_slice_upper

Encodes input as a hex string using uppercase characters into a mutable slice of bytes output.

encode_upper

Encodes data as a hex string using uppercase characters.

encode_upper_prefixed

Encodes data as a prefixed hex string using uppercase characters.