Trait ethers::core::k256::elliptic_curve::ops::ReduceNonZero

pub trait ReduceNonZero<Uint>: Reduce<Uint> + Sizedwhere
    Uint: Integer,{
    // Required methods
    fn reduce_nonzero(n: Uint) -> Self;
    fn reduce_nonzero_bytes(bytes: &Self::Bytes) -> Self;
}

Modular reduction to a non-zero output.

This trait is primarily intended for use by curve implementations such as the k256 and p256 crates.

End users should use the Reduce impl on NonZeroScalar instead.

Required Methods

fn reduce_nonzero(n: Uint) -> Self

Perform a modular reduction, returning a field element.

fn reduce_nonzero_bytes(bytes: &Self::Bytes) -> Self

Interpret the given bytes as an integer and perform a modular reduction to a non-zero output.