Module ethers::core::k256::elliptic_curve::point

Traits for elliptic curve points.

Structs

NonIdentity

Non-identity point type.

Traits

AffineCoordinates

Access to the affine coordinates of an elliptic curve point.

BatchNormalize

Normalize point(s) in projective representation by converting them to their affine ones.

CurveArithmetic

Elliptic curve with an arithmetic implementation.

DecompactPoint

Decompact an elliptic curve point from an x-coordinate.

DecompressPoint

Decompress an elliptic curve point.

Double

Double a point (i.e. add it to itself)

PointCompaction

Point compaction settings.

PointCompression

Point compression settings.

Type Aliases

AffinePoint

Affine point type for a given curve with a CurveArithmetic implementation.

ProjectivePoint

Projective point type for a given curve with a CurveArithmetic implementation.