Module ethers::middleware::transformer::ds_proxy::factory

Modules

ds_proxy_factory

This module was auto-generated with ethers-rs Abigen. More information at: https://github.com/gakonst/ethers-rs

Structs

BuildCall

Container type for all input parameters for the build function with signature build(address) and selector 0xf3701da2

BuildReturn

Container type for all return fields from the build function with signature build(address) and selector 0xf3701da2

BuildWithSenderCall

Container type for all input parameters for the build function with signature build() and selector 0x8e1a55fc

BuildWithSenderReturn

Container type for all return fields from the build function with signature build() and selector 0x8e1a55fc

CacheCall

Container type for all input parameters for the cache function with signature cache() and selector 0x60c7d295

CacheReturn

Container type for all return fields from the cache function with signature cache() and selector 0x60c7d295

CreatedFilter

DsProxyFactory

IsProxyCall

Container type for all input parameters for the isProxy function with signature isProxy(address) and selector 0x29710388

IsProxyReturn

Container type for all return fields from the isProxy function with signature isProxy(address) and selector 0x29710388

Enums

DsProxyFactoryCalls

Container type for all of the contract’s call

Statics

ADDRESS_BOOK

A lazily computed hash map with the Ethereum network IDs as keys and the corresponding DsProxyFactory contract addresses as values

DSPROXYFACTORY_ABI

The parsed JSON ABI of the contract.

DSPROXYFACTORY_BYTECODE

The bytecode of the contract.